update for for fields in the VCAP service JSON
